sentinel_x/requirements.txt
flask>=3.0.0
pymongo>=4.6.0
pandas>=2.1.0
scikit-learn>=1.3.0
azure-monitor-query>=1.2.0
azure-identity>=1.15.0
kubernetes>=28.1.0
requests>=2.31.0
python-dotenv>=1.0.0
